New reactive movement code
Finally something that can be understood by humans (and much faster than
what I had). I removed all the ridiculous unnecessary vector math and
went for a simple "target angle" concept. This involved getting rid of
the Vector in Nerves - a good idea on paper that generated ridiculous
amounts of pain. The new code still needs a cleanup, and the calculation
of motion forces still sucks - but at least now I have something that
makes sense to my mind.

While I was here, I also added a new IsolationChamber to play with a
single narjillo and analyze its movements. In the process, I found that
the Pond class is extremely fragile in the presence of concurrent
accesses. Sigh.