Remove constraints on lateral movement
While rewriting the physics code, I apparently managed to eliminate
swarmers much earlier than I hoped. :) I was assuming that removing
swarmers would have required an entirely new physics engine, to avoid
creatures with a small head that barely moves, and a big tail that
oscillates wildly. Actually, the problem with swarmers seemed to be
related to the way I managed angles. Creatures still have an
"impossible" angle setup (the "tail wiggling the dog" effect), but this
doesn't seem to be a huge problem. I can probably live with it.