Remove fast sin and cos tables
They did improve speed, but the minimal improvement (5-10%) simply
wasn't worth worrying about the lack in precision. The lack in precision
did have an effect, judging from the very different results of
experiments. It's hard to tell whether it was worse or simply different,
but I decided not to take the risk and simply revert to Java's standard
Math.sin() and Math.cos(). I'll consider re-adding fast trigonometric
tables in the future, if the next iterations of the physics engine
happen to do a lot of trigonometry.