More deep refactoring and fix rotation bug
This is a biggie. Apparently, the forces were calculated based on a
relative center of mass, not the absolute one. As a result, rotation was
completely b0rked. Now that I fixed this, all the creatures' rotational
movements appear much more natural. Next step is applying some real
physics to calculate rotations instead of the current ad-hoc code.