Fix vector math
I've been too quick in assuming "edge conditions" are not a problem.
With vector math, they are: all operations that rely on vector angles
are potentially vulnerable to zero-length vectors, and this is actually
a big deal. I used checked exceptions for once - it's important that
every caller is checking the parameters when doing any
operation with vectors that might involve a zero-length vector.