Fix bugs in center of mass calculation
First of all, it was calculated erroneously for each organ. To boot,
there was a hidden cast to int in the main calculation that might also
have resulted in tiny errors. The first problem was huge: essentially,
the center of mass was completely wrong (shifted much closer to the head
than it should have been - in large creatures, this resulted in a big
error).