Fix bug in linear momentum
I was considering any movement of each body segment in any direction.
Instead, only the normal component of the movement should count. A body
segment doesn't generate any push if it slides along its axis.

This code also contains a first stab at implementing a bonus for being
fast, in the form of a viscosity factor (on the whole body only).