Fix bug with miscalculated moments of inertia
This one was sneaky, and the main culprit of the remainins "weird"
movements. The problem: when I calculated movement for body segments, I
approximated it as the movement of the center point. I thought this
would be acceptable, but it ended up creating very wrong movements for
segments that were rotating. I reverted to my old style of calculating
movement by averaging the translation of the two endpoints.