Fix bug with non-deterministic experiments
This one took me hours. The cause of non-deterministic behaviour was
that I used HashMaps and Sets, that have no guarantee on ordering.
Everynow and then, in particular when checking collisions with food, the
system would check creatures/food in a different order, and this was
enough to cause a butterfly effect through the simulation. I moved to
using only ordered LinkedHashSets and LinkedHashMaps. To make things
worse, the bug wasn't usually visible in Experiment - only in PetriDish,
possibly because of some non-deterministic interaction between classloaders,
threads and cosmic rays.