Use space partitioning to check collisions
This is an optimization that was long overdue. Instead of using one
large space to look for collisions between Narjillos and food, I
partition the space into a grid of small areas. As long as a Narjillo
never moves fast enough to skip an area, I can search for food only in
the neighboring squares of the Narjillo. This makes the entire system
almost 30% faster.