Realistic calculation of creature radius
Earlier on, the moment of inertia of a narjillo always considered its
maximum theoretical radius. Now I used the real instantaneous radius.
This takes a few extra calculations, but it also allows creatures to
spin easier by "contracting" their body shape. This comes at a pretty
high performance cost, but it's probably worth it for the sake of
realistic physics.