Persistent Experiment
This was a lot of work, but it was worth it: now every run of both
"experiment" and "petri" writes a file with the entire state of the
experiment (updated every 10K ticks, right before printing out
the status). If you pass this file as a parameter to "experiment" or
"petri", the experiment will pick up from the latest 10K tick mark.
This allows very long experiments, without any fear of losing data.