Make chromosome mutations much rarer
So far, adding or removing a chromosome was as common as mutating a
gene, but the consequences were often much larger. I made chromosome
mutations rarer than gene mutations, by a factor equal to the number of
genes in a chromosome. This slows down change, so I raised the chances
of mutations in general.