Fix bug with body plan tree structure
The body tree contained the wrong relationships between parents and
children. As a result, the same body segment was being added multiple
times. This explained both the sudden drop in performance (I had
thousands of unnecessary body part physics calculations) and the failure
of the "non-deterministic experiment" test (which had to do with the
fact that the same object was being serialized multiple times, and
was then deserialized as multiple equal but not identical objects).