Increase energy cost exponentially for higher metabolic rates.
Metabolic rates used to be boring, because there was no reason to have a
low rate - you just moved slower, at the same energy cost per distance.
As a result, all narjillos evolved to have the highest possible
metabolic rate. Now energy consumption increases with the metabolic rate
to the power of 1.5, which makes for a more interesting trade-off.