Add histeresis to target direction
The creatures' mouths now cannot point in the opposite direction of
their head. This hopefully fixes the problem with creatures pointing
their mouth backwards, but moving in the same way they'd move if the
food was straight in front of them. This last issue is a side effect of
the way skewing works: if food is behind you, then you have no skewing
(as opposed to maximum skewing, which you get when food is to your
side).