Fix bug that caused non-deterministic experiments over the long run
A narjillo had a chance to consume food before another narjillo had a
chance to target it. This caused a very small and very rare error - but
it did happen over tens of thousands of ticks, and it could have a
butterfly effect on evolution. The standard DeterministicExperimentTest
didn't catch this, because it didn't tick for quite long enough for it
to happen reliably.