Fix concurrent modification bug
For a description of the effect of the bug, see the backlog in the
previous commit. The problem is that Java's "concurrent-ready"
collections actually aren't - when iterating over a collection, the
entire collection isn't locked (which does make some sense). This means
that I could still get ConcurrentModificationExceptions. I could have
used forEach(), that apparently is thread-safe, but this is somewhat at
odds with the current Java documentation (see:
http://stackoverflow.com/questions/23450227/is-iteration-via-collections-synchronizedset-foreach-guaranteed-to-be-thr).
I just decided to make things more explicit and synchronize by hand.