Quadruple the size of an experiment
The ecosystem is now four times as large (with four times as many eggs
and narjillos), to make the system more stable. Also, food respawns four
times faster, to support a larger population. As a pleasant side effect,
the performance test is now more in line with the reality of me owning a
brand new quad-core machine. :)