Migrate from Gradle to Buildr
After using Gradle for a while, I came to the conclusion that I'd rather
use anything else. It's *that* bad. Buildr is a bit painful during
installation (it requires Ruby), but then it's remarkably intuitive. The
current buildscript is still not 100% usable, but it's much better than
the Gralde build ever was.

I also took the chance to put the two subprojects together into one
large project. They were originally separated because I wanted a chance
to run experiments without installing JavaFX. JavaFX is now a standard
library in Java 8, so there is no need to keep the graphics stuff
separate anymore, since I migrated even the core code to Java 8 anyway.