Replace Levenshtein distance with SimHash distance
I introduced the SimHash algorithm to easy identify species, but it
works well for checking distance as well - so I switched to this one
instead of the simpler Levenshtein distance I've been using this far.

To be extra picky, Levenshtein's distance is technically a better
indicator of genetic distance than SimHash, because it's a tad more
semantic: it looks at gene distance rather than binary distance. (So,
for example, a gene of value 2 is correctly assumed to be similar to a
gene of value 1, while SimHash looks at the binary repesentations and
sees 0b01 as quite different from 0b10, but more similar to 0b11).
However, I was using Levenshtein distances only for a marginal use case
(identifying the "most successfull" genotype in the pool), while SimHash
is useful for coarse identification of species with a simple 24-bit
hash. I'll probably use SimHash soon, and I'm planning to switch to a
much more powerful clustering algorithm to identify species. So
I decided that SimHas is Good Enough as an indicator of genetic distance
for now, and the Levenshtein distance algorithm isn't worth the
additional code.