Export ancestry to CSV
For use in programs such as Gephi. This is more useful than a NEXUS
tree, because Gephi supports it, and Gephi seems to be one rare case of
a program that can handle large trees. Still, it cannot handle anything
as large as what we get in a long, multi-Mticks experiment. I should
consider doing away with NEXUS, and possibly even CSV, if I find out
that I cannot extract useful data from them.