Move serial number generation to NumGen
The class was previously known as RanGen, but now it can generate serial
numbers besides random numbers. It's an ideal place to put the serial
number, because the number will be saved together with the rest of the
Experiment (which carries the NumGen along for the ride). This means
that we don't need to peek into the database to find the next available
serial, and as a bonus, serial number generation (which is almost
global, hence potentially prone to race conditions) will be checked for
illegal access by multiple threads, which is another feature of the
NumGen. I'm pretty proud of this idea. It looks stupidly obvious in
hindsight, so it must be good.