New -G option in lab program
Counts the number of completely separate germlines in the pond - that
is, it clusters creatures that have a common ancestor, and then counts
the clusters. This proves that actual speciation is happening: in long
experiments, even creatures that have separate echological niches are
actually descendants of a common ancestor. (The result of running "lab
-G" is 1).